backend/requirements.txt
Flask==3.0.0
pymongo==4.5.0
pandas==2.1.2
scikit-learn==1.3.2
flask-cors==4.0.0
google-genai
python-dotenv
requests
